Determine whether the user enter the true answer
Create a while loop to check whether the user enter the true answer, if the user enter the wrong answer then into the loop and store the wrong answer in the arraylist, after that, check whether the wrong answer is enter the same before, then ask the user enter the answer again.  If the user enter the true answer, then print the message to tell the user that the answer is true.